+++ flask_dance/consumer/backend/__init__.py
@@ -1,56 +1,23 @@
-import six
-from abc import ABCMeta, abstractmethod
+import warnings
+from flask_dance.consumer.storage import BaseStorage as BaseBackend
+from flask_dance.consumer.storage import NullStorage, MemoryStorage
 
 
-class BaseBackend(six.with_metaclass(ABCMeta)):
-    @abstractmethod
-    def get(self, blueprint):
-        return None
-
-    @abstractmethod
-    def set(self, blueprint, token):
-        return None
-
-    @abstractmethod
-    def delete(self, blueprint):
-        return None
+class NullBackend(NullStorage):
+    def __init__(self):
+        warnings.warn(
+            "flask_dance.consumer.backend.NullBackend is deprecated. "
+            "Please use flask_dance.consumer.storage.NullStorage instead.",
+            DeprecationWarning,
+        )
+        super(NullBackend, self).__init__()
 
 
-class NullBackend(BaseBackend):
-    """
-    This mock storage will never store OAuth tokens.
-    If you try to retrieve a token from this storage, you will always
-    get ``None``.
-    """
-
-    def get(self, blueprint):
-        return None
-
-    def set(self, blueprint, token):
-        return None
-
-    def delete(self, blueprint):
-        return None
-
-
-class MemoryBackend(BaseBackend):
-    """
-    This mock storage stores an OAuth token in memory and so that it can
-    be retrieved later. Since the token is not persisted in any way,
-    this is mostly useful for writing automated tests.
-
-    The initializer accepts a ``token`` argument, for setting the
-    initial value of the token.
-    """
-
+class MemoryBackend(MemoryStorage):
     def __init__(self, token=None, *args, **kwargs):
-        self.token = token
-
-    def get(self, blueprint):
-        return self.token
-
-    def set(self, blueprint, token):
-        self.token = token
-
-    def delete(self, blueprint):
-        self.token = None
+        warnings.warn(
+            "flask_dance.consumer.backend.MemoryBackend is deprecated. "
+            "Please use flask_dance.consumer.storage.MemoryStorage instead.",
+            DeprecationWarning,
+        )
+        super(MemoryBackend, self).__init__(token, *args, **kwargs)
